JBSFRAME-37 Refactorizado DataQueryModel Mejorando documentación
